refactor ttu, new folder for integrations
